Move Euclidean geometry to 'euclidean' submodule. Create small circle method.
