+++ plot_tools/euclidean/__init__.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+#
+# Plot tools euclidean geometry submodule init file.
+# Copyright (C) 2018 Malte Ziebarth
+# 
+# This software is distributed under the MIT license.
+# See the LICENSE file in this repository.
+
+from .euclidean import *
